adding matrix converter to that branch

+++ include/ngchol/datatypes.hpp
@@ -10,10 +10,12 @@
 
 namespace LIBCHOLESKY {
 
-typedef    long long int              LongInt;
+//typedef    long long int              LongInt;
+typedef    int              LongInt;
 typedef    int                   Int;
 typedef    uint32_t              Idx32;
-typedef    uint64_t              Idx64;
+typedef    uint32_t              Idx64;
+//typedef    uint64_t              Idx64;
 typedef    double                Real;
 typedef    std::complex<double>  Complex; // Must use elemental form of complex
 #ifdef _USE_COMPLEX_
